mobile and web card

+++ users-mobile/src/api/api.js
@@ -29,11 +29,12 @@ api.interceptors.request.use(async (config) => {
 api.interceptors.response.use(
   (response) => response,
   async (error) => {
-    if (error.response?.status === 401) {
+    if (error.response?.status === 401 && !error.config?.skipUnauthorizedHandler) {
       await clearHceToken();
       await SecureStore.deleteItemAsync('token');
       await SecureStore.deleteItemAsync('passengerName');
       await SecureStore.deleteItemAsync('tempToken');
+      await SecureStore.deleteItemAsync('passengerCardNumber');
       await clearHceToken();
 
       if (unauthorizedHandler) {
